Getting there, but profile still isn't giving us the correct tracer
